split session routes
